More Search page improvements (#119)
* Hooked up search results for season, category and streaming.
Fixed navigation in FilterSub.
Added keyExtractor to ContentList.

* Hooked up search results and media pages.
Added CustomHeader to media pages when it's loading. A user should be able to navigate back without having to wait for the page to load incase they accidentally tapped the media.

* Improved Results layout.
This is temporary however, this will be made so it tries to get the best fitting of cards in the results so that devices with bigger screens don't get a lot of blank space in between cells.

* Made Grid look much better in ResultsView.

Added getGridItemSpacing function to utils. This will calculate margins necessary to get a uniform grid look across multiple different devices/screen sizes.

* Quick image size revert.

* Changed to ViewPropTypes for style.

* Quick fixes.

* Fixed up docs.